uws: shorten stream-buffer comments, move the source lint to source-lints
The lint reads only src/uws_sys/us_socket_t.rs and never runs the built
binary, so it belongs in test/internal/source-lints/ where the GitHub
Actions lane runs it against a released bun. Its header now cites the
miri job in rust-lints.yml instead of the removed miri.yml.
