+++ src/runtime/socket/uws_jsc.rs
@@ -141,12 +141,8 @@ unsafe extern "C" fn us_socket_buffered_js_write(
     // `&mut *socket` / `&mut *buffer` under Stacked Borrows, so raw pointers with
     // no uniqueness assertion are used throughout.
 
-    // Convert `data`/`encoding` BEFORE taking the stream buffer: the conversion can
-    // run arbitrary JS (toString/Symbol.toPrimitive, Request/Response body coercion)
-    // which can re-enter this function on the same socket. A re-entrant call while
-    // this frame holds the taken buffer would take an empty buffer, stash its own
-    // data there, and the final `update` below would drop that data and reorder the
-    // stream.
+    // Coerce `data` before taking the buffer: the coercion can run JS that
+    // re-enters this function on the same socket.
     let node_buffer: BlobOrStringOrBuffer = if data.is_undefined() {
         BlobOrStringOrBuffer::StringOrBuffer(StringOrBuffer::EMPTY)
     } else {
@@ -180,9 +176,8 @@ unsafe extern "C" fn us_socket_buffered_js_write(
         }
     }
 
-    // SAFETY: caller (JSNodeHTTPServerSocket.cpp) guarantees `buffer` is valid for the call.
-    // The take nulls the raw parts, so this owning `Vec` is the allocation's sole owner;
-    // no JS executes between here and the `update()` below (see the ordering note above).
+    // SAFETY: the C++ caller keeps `buffer` valid for the call. No JS runs
+    // between here and the `update()` below.
     let mut stream_buffer = unsafe { &mut *buffer }.take_stream_buffer();
     let mut total_written: usize = 0;
